utils/pyproject.toml
[tool.poetry]
name = "faust-utils"
version = "0.1.0"
description = "Utility scripts for the Faust Edition"
authors = ["Thorsten Vitt <thorsten.vitt@uni-wuerzburg.de>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
lxml = "^4.6.3"
treetaggerwrapper = "^2.3"
pandas = "^1.3.0"
tqdm = "^4.61.2"
ftfy = "^6.0.3"
scikit-image = "^0.18.2"
joblib = "^1.0.1"
ipykernel = "^6.0.3"
openpyxl = "^3.0.7"
"ruamel.yaml" = "^0.17.10"
Pillow = "^9.1.0"
rich = "^12.4.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
rope = "^0.19.0"
lxml-stubs = "^0.4.0"
line-profiler-pycharm = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
